requirements.txt
pytest==8.2.1
pytest-playwright==0.7.0
allure-pytest==2.14.3